Split arrow_expression.rs into a dedicated submodule
Move tests and utility functions to separate files for better organization and maintainability.
